+++ src/internal/handler/mod.rs
@@ -1,2 +1,3 @@
 pub mod user_handler;
 pub mod company_handler;
+
