+++ src/internal/handler/user_handler/mod.rs
@@ -1,0 +1,4 @@
+pub mod sign_up_handler;
+pub mod log_in_handler;
+pub mod change_user_handler;
+pub mod get_user_handler;
