+++ src/router/user_router.rs
@@ -2,7 +2,12 @@ use axum::{
     routing::{get, post, put}, Router
 };
 
-use crate::internal::handler::user_handler::UserHandler;
+use crate::internal::handler::user_handler::{
+    sign_up_handler::SignUpHandler,
+    log_in_handler::LogInHandler,
+    change_user_handler::ChangeUserHandler,
+    get_user_handler::GetUserHandler,
+};
 
 pub fn create_router() -> Router {
     // 공유 레포지토리 생성
@@ -11,9 +16,9 @@ pub fn create_router() -> Router {
         .nest(
             "/user",
             Router::new()
-                .route("/sign-up", post(UserHandler::sign_up_handler))
-                .route("/log-in", post(UserHandler::log_in_handler))
-                .route("/get-user", get(UserHandler::get_user_handler))
-                .route("/change-user", put(UserHandler::change_user_handler))
+                .route("/sign-up", post(SignUpHandler::sign_up_handler))
+                .route("/log-in", post(LogInHandler::log_in_handler))
+                .route("/get-user", get(GetUserHandler::get_user_handler))
+                .route("/change-user", put(ChangeUserHandler::change_user_handler))
         )
 }
